Allow doubles in incoming sensor data, to prepare for sds011.

+++ workspace/samenmetenbridge/src/main/java/nl/sikken/bertrik/sensor/SensorPmTriplet.java
@@ -10,13 +10,13 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public final class SensorPmTriplet {
 
     @JsonProperty("pm1_0")
-    private Integer pm1_0;
+    private Double pm1_0;
     
     @JsonProperty("pm2_5")
-    private Integer pm2_5;
+    private Double pm2_5;
     
     @JsonProperty("pm10")
-    private Integer pm10;
+    private Double pm10;
     
     private SensorPmTriplet() {
         // Jackson constructor
@@ -29,22 +29,22 @@ public final class SensorPmTriplet {
      * @param pm2_5 PM2.5 concentration in ug/m3
      * @param pm10 PM10 concentration in ug/m3
      */
-    public SensorPmTriplet(Integer pm1_0, Integer pm2_5, Integer pm10) {
+    public SensorPmTriplet(Double pm1_0, Double pm2_5, Double pm10) {
         this();
         this.pm1_0 = pm1_0;
         this.pm2_5 = pm2_5;
         this.pm10 = pm10;
     }
     
-    public Integer getPm1_0() {
+    public Double getPm1_0() {
         return pm1_0;
     }
 
-    public Integer getPm2_5() {
+    public Double getPm2_5() {
         return pm2_5;
     }
 
-    public Integer getPm10() {
+    public Double getPm10() {
         return pm10;
     }
 
@@ -52,7 +52,7 @@ public final class SensorPmTriplet {
      * @see java.lang.Object#toString()
      */
     public String toString() {
-        return String.format(Locale.US, "{pm1_0=%d,pm2_5=%d,pm10=%d}", pm1_0, pm2_5, pm10);
+        return String.format(Locale.US, "{pm1_0=%.1f,pm2_5=%.1f,pm10=%.1f}", pm1_0, pm2_5, pm10);
     }
     
 }
